refactor: i18n 관련 설정 파일 통합
