+++ src/app/[lang]/(news)/news/page.tsx
@@ -1,6 +1,6 @@
 import News from "@/app/[lang]/(news)/news/_components";
 import { getDictionary } from "@/get-dictionary";
-import { Locale } from "@/i18n-config";
+import { Locale } from "@/i18n";
 
 export default async function NewsPage({ params: { lang } }: { params: { lang: Locale } }) {
   const dictionary = await getDictionary(lang);
