+++ src/get-dictionary.ts
@@ -1,5 +1,5 @@
 import "server-only";
-import type { Locale } from "./i18n-config";
+import type { Locale } from "./i18n";
 
 // We enumerate all dictionaries here for better linting and typescript support
 // We also get the default import for cleaner types
