--- src/i18n-config.ts
@@ -1,6 +1,0 @@
-export const i18n = {
-  defaultLocale: "ko",
-  locales: ["ko", "en", "ch", "jp", "fr"],
-} as const;
-
-export type Locale = (typeof i18n)["locales"][number];
