+++ src/middleware.ts
@@ -1,6 +1,6 @@
 import { match } from "path-to-regexp";
 import { NextRequest, NextResponse } from "next/server";
-import { i18n } from "./i18n-config";
+import { i18n } from "./i18n";
 import { match as matchLocale } from "@formatjs/intl-localematcher";
 import Negotiator from "negotiator";
 import { getLanguageCookie } from "./utils/cookies";
